Extract collapse contracts into named hooks (#276)
Pure refactor: the three collapse contracts (broadcast application,
prop-change retire, user-action clear) were interleaved across
CollapseProvider, CollectionNode, and ValueNodeWrapper without their
boundaries being visible. Each is now named so the call sites read
top-to-bottom and §16 can replace the per-node broadcast effect with a
per-path atomic store without disturbing them.

No behavior changes, no API changes; 232 tests pass unchanged.
