+++ src/contexts/index.ts
@@ -1,4 +1,4 @@
 export * from './ThemeProvider'
 export { TreeStateProvider } from './TreeStateProvider'
 export { useEditing } from './EditingProvider'
-export { useCollapse, doesCollapseStateMatchPath } from './CollapseProvider'
+export { useCollapse, useAppliedBroadcast, useReferenceChanged } from './CollapseProvider'
